+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+# 0.1.1
+* Fix special encode/decode rule for Option\<Bool\>.
+* Fix encode/decode for signed integers to avoid overflow.
+* Add missing u256.
+* Wrap all exceptions to Exception object.
+* Add test cases for base data types.
+* Follow dart code conventions.
+
 # 0.1.0
 * Initial release